+++ exonum-java-binding/common/src/main/java/com/exonum/binding/common/transaction/TransactionLocation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 The Exonum Team
+ * Copyright 2019 The Exonum Team
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,9 +12,10 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package com.exonum.binding.blockchain;
+package com.exonum.binding.common.transaction;
 
 import com.google.auto.value.AutoValue;
 import com.google.gson.Gson;
