+++ exonum-java-binding/core/src/main/java/com/exonum/binding/blockchain/serialization/TransactionLocationSerializer.java
@@ -18,8 +18,8 @@ package com.exonum.binding.blockchain.serialization;
 
 import static com.exonum.binding.common.serialization.StandardSerializers.protobuf;
 
-import com.exonum.binding.blockchain.TransactionLocation;
 import com.exonum.binding.common.serialization.Serializer;
+import com.exonum.binding.common.transaction.TransactionLocation;
 
 public enum TransactionLocationSerializer implements Serializer<TransactionLocation> {
   INSTANCE;
