+++ exonum-java-binding/core/src/test/java/com/exonum/binding/blockchain/TransactionResultTest.java
@@ -21,7 +21,8 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.exonum.binding.blockchain.TransactionResult.Type;
+import com.exonum.binding.common.transaction.TransactionResult;
+import com.exonum.binding.common.transaction.TransactionResult.Type;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
