+++ exonum-java-binding/core/src/test/java/com/exonum/binding/blockchain/serialization/TransactionResultSerializerTest.java
@@ -19,8 +19,8 @@ package com.exonum.binding.blockchain.serialization;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-import com.exonum.binding.blockchain.TransactionResult;
 import com.exonum.binding.common.serialization.Serializer;
+import com.exonum.binding.common.transaction.TransactionResult;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
