+++ exonum-java-binding/qa-service/src/test/java/com/exonum/binding/qaservice/ApiControllerIntegrationTest.java
@@ -56,11 +56,8 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.exonum.binding.blockchain.Block;
-import com.exonum.binding.blockchain.TransactionLocation;
-import com.exonum.binding.blockchain.TransactionResult;
 import com.exonum.binding.blockchain.serialization.BlockAdapterFactory;
 import com.exonum.binding.blockchain.serialization.TransactionLocationAdapterFactory;
-import com.exonum.binding.blockchain.serialization.TransactionResultAdapterFactory;
 import com.exonum.binding.common.configuration.ConsensusConfiguration;
 import com.exonum.binding.common.configuration.StoredConfiguration;
 import com.exonum.binding.common.configuration.ValidatorKey;
@@ -73,6 +70,9 @@ import com.exonum.binding.common.hash.HashFunction;
 import com.exonum.binding.common.hash.Hashing;
 import com.exonum.binding.common.message.TransactionMessage;
 import com.exonum.binding.common.serialization.json.JsonSerializer;
+import com.exonum.binding.common.serialization.json.TransactionResultAdapterFactory;
+import com.exonum.binding.common.transaction.TransactionLocation;
+import com.exonum.binding.common.transaction.TransactionResult;
 import com.exonum.binding.service.InternalServerError;
 import com.exonum.binding.test.Bytes;
 import com.google.common.collect.ImmutableMap;
